requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
boto3==1.24.26
botocore==1.27.26
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.5
dj-database-url==0.5.0
Django==4.0.6
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-jazzmin==2.5.0
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
django-storages==1.12.3
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djoser==2.1.0
drf-yasg==1.21.3
filelock==3.7.1
gunicorn==20.1.0
idna==3.3
importlib==1.0.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
mailjet-rest==1.3.4
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
phonenumbers==8.12.53
Pillow==9.2.0
platformdirs==2.5.2
portalocker==2.5.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
pywin32==304
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.10
virtualenv==20.16.2
waitress==2.1.2
whitenoise==6.2.0
wrapt==1.14.1